ahtml: commented idea of a `flat` method without implementation idea
Needs some real future, not sure where. Wish to keep translarent
elements out of the proper tree, because it would make recursing it
more painful. Thus needs to be some type that is automatically
flattened in when creating an element. That would have to be
represented via AId, since that is what we use to pass bodies in, and
it would still have to be AId<Node> to fit, too, and creation of
elements then would have to scan the bodies and flatten them. Which is
an additional cost, especially since they have to be dereferenced. Is
it worth it?